api/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.4
dataclasses-json==0.6.1
faiss-cpu==1.7.4
falcon==3.1.1
falcon-jwt-checker==0.1.3
falcon-multipart==0.2.0
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
huggingface-hub==0.19.4
idna==3.4
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.323
langsmith==0.0.52
MarkupSafe==2.1.3
marshmallow==3.20.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.1
openai==0.28.1
packaging==23.2
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyjson==1.4.1
PyJWT==2.8.0
pypdf==3.16.4
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
sniffio==1.3.0
SQLAlchemy==2.0.22
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.1
tokenizers==0.15.0
torch==2.1.1
tqdm==4.66.1
transformers==4.35.2
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.0
yarl==1.9.2
